Bi-weekly merge

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@40234 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/blast/seqdb_writer/unit_test/writedb_unit_test.cpp
@@ -2028,7 +2028,7 @@ BOOST_AUTO_TEST_CASE(TooManyAlgoId)
     // Ensure that the IDs were assigned in increasing order
     CHECK_EQUAL(kMaxNumSupportedAlgorithmVariants, algo_ids.size());
     for (size_t i = 0; i < algo_ids.size(); i++) {
-        CHECK_EQUAL((masking_algorithm + i), algo_ids[i]);
+        CHECK_EQUAL((int)(masking_algorithm + i), (int)algo_ids[i]);
     }
 }
